backend/pyproject.toml
[tool.poetry]
name = "brk-13f-backend"
version = "0.1.0"
description = "Berkshire 13F ingestion and API"
authors = []
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.32.0" }
sqlmodel = "^0.0.22"
pydantic = "^2.0"
pydantic-settings = "^2.0"
httpx = "^0.28.0"
pandas = "^2.2.0"
lxml = "^5.0.0"
apscheduler = "^3.10.0"
structlog = "^24.0.0"
alembic = "^1.14.0"
python-dotenv = "^1.0.0"
aiosqlite = "^0.20.0"
greenlet = "^3.0.0"
google-auth = "^2.0"
google-api-python-client = "^2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
black = "^24.0.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.13.0"
pre-commit = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
